backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
biopython>=1.84
openai>=1.30.0
httpx>=0.27.0
pyzotero>=1.5.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
